Have the PySyft controller create its own identity
We were unable to reconnect to the unity backend with the same worker id
due to how zmq handles its sockets internally.

If the unity backend was restarted, and the client tried to remake the
controller (which you think would give you a new socket, thus acting
like a brand new session) but used the same identifier, then zmq would
very often reuse the same socket.

With more work we would be able to reuse the same socket.  I will add
that in a later PR but for now this is much better and should give
developer productivity a small boost